+++ src/main/java/nbbang/com/nbbang/domain/hashtag/domain/Hashtag.java
@@ -1,12 +1,10 @@
-package nbbang.com.nbbang.domain.party.domain;
+package nbbang.com.nbbang.domain.hashtag.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 
 import javax.persistence.*;
-import java.util.List;
 
 @Entity @Getter @Builder
 @AllArgsConstructor
