+++ src/main/java/nbbang/com/nbbang/domain/hashtag/domain/PartyHashtag.java
@@ -1,8 +1,9 @@
-package nbbang.com.nbbang.domain.party.domain;
+package nbbang.com.nbbang.domain.hashtag.domain;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
+import nbbang.com.nbbang.domain.party.domain.Party;
 
 import javax.persistence.*;
