+++ src/main/java/nbbang/com/nbbang/domain/hashtag/repository/HashtagRepository.java
@@ -1,0 +1,10 @@
+package nbbang.com.nbbang.domain.hashtag.repository;
+
+import nbbang.com.nbbang.domain.hashtag.domain.Hashtag;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface HashtagRepository extends JpaRepository<Hashtag, Long> {
+    List<Hashtag> findByContent(String content);
+}
