+++ src/main/java/nbbang/com/nbbang/domain/party/dto/PartyFindResponseDto.java
@@ -6,7 +6,6 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 import nbbang.com.nbbang.domain.member.domain.Member;
 import nbbang.com.nbbang.domain.party.domain.Party;
-import nbbang.com.nbbang.domain.party.domain.PartyHashtag;
 
 import java.time.LocalDateTime;
 import java.util.List;
