--- src/main/java/nbbang/com/nbbang/domain/party/repository/HashtagRepository.java
@@ -1,12 +1,0 @@
-package nbbang.com.nbbang.domain.party.repository;
-
-import nbbang.com.nbbang.domain.party.domain.Hashtag;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
-import java.util.List;
-
-public interface HashtagRepository extends JpaRepository<Hashtag, Long> {
-    List<Hashtag> findByContent(String content);
-}
