+++ src/test/java/nbbang/com/nbbang/domain/party/service/ManyPartyServiceTest.java
@@ -3,13 +3,13 @@ package nbbang.com.nbbang.domain.party.service;
 import nbbang.com.nbbang.domain.member.domain.Member;
 import nbbang.com.nbbang.domain.member.dto.Place;
 import nbbang.com.nbbang.domain.member.repository.MemberRepository;
-import nbbang.com.nbbang.domain.party.domain.Hashtag;
+import nbbang.com.nbbang.domain.hashtag.domain.Hashtag;
 import nbbang.com.nbbang.domain.party.domain.Party;
-import nbbang.com.nbbang.domain.party.domain.PartyHashtag;
+import nbbang.com.nbbang.domain.hashtag.domain.PartyHashtag;
 import nbbang.com.nbbang.domain.party.domain.PartyStatus;
 import nbbang.com.nbbang.domain.party.dto.many.PartyListRequestFilterDto;
-import nbbang.com.nbbang.domain.party.repository.HashtagRepository;
-import nbbang.com.nbbang.domain.party.repository.PartyHashtagRepository;
+import nbbang.com.nbbang.domain.hashtag.repository.HashtagRepository;
+import nbbang.com.nbbang.domain.hashtag.repository.PartyHashtagRepository;
 import nbbang.com.nbbang.domain.party.repository.PartyRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
